Refactor - extract Appointment type to common types
