+++ imports/common/types/Appointment.ts
@@ -1,0 +1,7 @@
+export type Appointment = {
+  _id: string
+  firstName: string
+  lastName: string
+  date: Date
+  userId: string
+}
