+++ imports/db/Appointments.ts
@@ -1,11 +1,4 @@
 import { Mongo } from 'meteor/mongo'
-
-export type Appointment = {
-  _id: string
-  firstName: string
-  lastName: string
-  date: Date
-  userId: string
-}
+import { Appointment } from '../common/types/Appointment'
 
 export const Appointments = new Mongo.Collection<Appointment>('appointments')
